# apps/console-web/src/shared/api/types.ts
import type { components } from "./generated/control-plane";

export type Session = components["schemas"]["Session"];
export type MFAChallenge = components["schemas"]["MFAChallenge"];
export type MFAStatus = components["schemas"]["MFAStatus"];
export type TOTPEnrollment = components["schemas"]["TOTPEnrollment"];
export type User = components["schemas"]["User"];
export type UserSession = components["schemas"]["UserSession"];
export type WorkspaceMembership = components["schemas"]["WorkspaceMembership"];
export type WorkspaceGroup = components["schemas"]["WorkspaceGroup"];
export type AuditEvent = components["schemas"]["AuditEvent"];
export type AccessGrant = components["schemas"]["AccessGrant"];
export type Workspace = components["schemas"]["Workspace"];
export type HierarchyInput = components["schemas"]["HierarchyInput"];
export type Project = components["schemas"]["Project"];
export type Environment = components["schemas"]["Environment"];
export type App = components["schemas"]["App"];
export type GitHubInstallation = components["schemas"]["GitHubInstallation"];
export type GitHubRepository = components["schemas"]["GitHubRepository"];
export type GitHubSource = components["schemas"]["GitHubSource"];
export type GitHubSourceInput = components["schemas"]["GitHubSourceInput"];
export type BuildInput = components["schemas"]["BuildInput"];
export type Build = components["schemas"]["Build"];
export type BuildLog = components["schemas"]["BuildLog"];
export type Release = components["schemas"]["Release"];
export type DeliveryPolicy = components["schemas"]["DeliveryPolicy"];
export type DeliveryPolicyInput = components["schemas"]["DeliveryPolicyInput"];
export type Variable = components["schemas"]["Variable"];
export type Parameter = components["schemas"]["Parameter"];
export type ParameterInput = components["schemas"]["ParameterInput"];
export type RuntimeConfiguration = components["schemas"]["RuntimeConfiguration"];
export type AppEnvironmentInput = components["schemas"]["AppEnvironmentInput"];
export type AppEnvironmentCreateInput = components["schemas"]["AppEnvironmentCreateInput"];
export type AppEnvironment = components["schemas"]["AppEnvironment"];
export type DeploymentInput = components["schemas"]["DeploymentInput"];
export type DeploymentPreviewInput = components["schemas"]["DeploymentPreviewInput"];
export type DeploymentPreview = components["schemas"]["DeploymentPreview"];
export type ConfigurationRevision = components["schemas"]["ConfigurationRevision"];
export type Deployment = components["schemas"]["Deployment"];
export type Operation = components["schemas"]["Operation"];
export type RuntimeLog = components["schemas"]["RuntimeLog"];
export type RuntimeLogs = components["schemas"]["RuntimeLogs"];
export type RuntimeLogBatch = components["schemas"]["RuntimeLogBatch"];
export type RuntimeMetricPoint = components["schemas"]["RuntimeMetricPoint"];
export type RuntimeMetricSeries = components["schemas"]["RuntimeMetricSeries"];
export type RuntimeMetrics = components["schemas"]["RuntimeMetrics"];
export type RuntimeMetricSample = components["schemas"]["RuntimeMetricSample"];
export type RuntimeMetricSnapshot = components["schemas"]["RuntimeMetricSnapshot"];
export type RuntimeEvent = components["schemas"]["RuntimeEvent"];
export type RuntimeEvents = components["schemas"]["RuntimeEvents"];
export type ApiError = components["schemas"]["Error"];
export type WorkspaceMutationAccepted = { workspace: Workspace; operation: Operation };
export type DeploymentMutationAccepted = { deployment: Deployment; operation: Operation };