# contracts/openapi/control-plane-v1.yaml
openapi: 3.1.0
info:
  title: Molejo Control Plane API
  version: 1.0.0
  description: Pre-alpha control-plane contract for the local beta workspace.
servers:
  - url: /
paths:
  /api/v1/session:
    post:
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '202':
          description: A second factor is required
          content: { application/json: { schema: { $ref: '#/components/schemas/MFAChallenge' } } }
    get:
      operationId: getSession
      responses:
        '200':
          description: Current session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '401': { $ref: '#/components/responses/Unauthorized' }
    delete:
      operationId: logout
      responses:
        '204': { description: Session revoked }
...
  /api/v1/workspaces/current:
    get:
      operationId: getCurrentWorkspace
      security: [{ sessionCookie: [] }]
      responses:
        '200':
          description: Authorized workspace
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Workspace' }
...
  /api/v1/operations/{operationId}:
    get:
      operationId: getOperation
      security: [{ sessionCookie: [] }]
      parameters:
        - name: operationId
          in: path
          required: true
          schema: { type: string, pattern: '^op-[a-z2-7]{20}$' }
      responses:
        '200':
          description: Operation
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Operation' }
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: fruto_session
...
  responses:
    Unauthorized:
      description: Authentication required
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
  schemas:
    LoginRequest:
      type: object
      required: [username, password]
      additionalProperties: false
      properties:
        username: { type: string, minLength: 3, maxLength: 64 }
        password: { type: string, minLength: 1, maxLength: 1024, writeOnly: true }
    Session:
      type: object
      required: [user, assuranceLevel, csrfToken, installationCapabilities, workspaceMemberships]
      additionalProperties: false
      properties:
        user: { $ref: '#/components/schemas/User' }
        assuranceLevel: { type: string, enum: [AAL1, AAL2] }
        csrfToken: { type: string, writeOnly: true }
        installationCapabilities:
          type: object
          required: [manageUsers, createWorkspace]
          additionalProperties: false
          properties:
            manageUsers: { type: boolean }
            createWorkspace: { type: boolean }
        workspaceMemberships:
          type: array
          items:
            type: object
            required: [workspaceId, role]
            additionalProperties: false
            properties:
              workspaceId: { type: string, pattern: '^ws-[a-z2-7]{20}$' }
              role: { type: string, enum: [Owner, Member, Viewer] }
...
    MFAChallenge:
      type: object
      required: [mfaRequired, method, challengeToken]
      additionalProperties: false
      properties:
        mfaRequired: { type: boolean, const: true }
        method: { type: string, enum: [TOTP] }
        challengeToken: { type: string, writeOnly: true }
...
    User:
      type: object
      required: [id, username, displayName, status, version, createdAt, updatedAt]
      additionalProperties: false
      properties:
        id: { type: string, pattern: '^usr-[a-z2-7]{20}$' }
        username: { type: string, minLength: 3, maxLength: 64 }
        displayName: { type: string, minLength: 1, maxLength: 120 }
        status: { type: string, enum: [Active, Disabled, Locked] }
        version: { type: integer, minimum: 1 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    Workspace:
      type: object
      required: [id, name, version, state, createdAt, updatedAt]
      additionalProperties: false
      properties:
        id: { type: string, pattern: '^ws-[a-z2-7]{20}$' }
        name: { type: string, minLength: 1, maxLength: 80 }
        version: { type: integer, minimum: 1 }
        state: { type: string, enum: [Pending, Running, Ready, Failed] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    Operation:
      type: object
      required: [id, kind, status, desiredVersion, attempts]
      properties:
        id: { type: string, pattern: '^op-' }
        appEnvironmentId: { type: string, pattern: '^aev-[a-z2-7]{20}$' }
        deploymentId: { type: string, pattern: '^dpl-[a-z2-7]{20}$' }
        kind: { type: string, enum: [ApplyDeployment, DeleteAppEnvironment, EnsureWorkspace] }
        status: { type: string, enum: [Pending, Running, Succeeded, Failed] }
        desiredVersion: { type: integer, minimum: 1 }
        attempts: { type: integer, minimum: 0 }
        errorCode: { type: string }
        errorMessage: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    Error:
      type: object
      required: [code, message, requestId]
      properties:
        code: { type: string }
        message: { type: string }
        requestId: { type: string }